// src/api/model/invoiceSettingsResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export interface InvoiceSettingsResponseDto {
  id: string;
  currency: string;
  invoicePrefix?: string;
  invoiceNumberFormat: string;
  nextInvoiceNumber: number;
  dueDaysDefault: number;
  defaultHeaderText?: string;
  defaultFooterText?: string;
  defaultVatRate?: number;
  logoUrl?: string;
  createdAt: string;
  updatedAt: string;
}